Split current node impls out into separate classes and rename to Simple*Node rather than *NodeImpl.
